Invert most of the logic in do_VN with fast returns.  Mostly flattens functions with clear logic now.

+++ src/vars.h
@@ -52,7 +52,7 @@ using commodtype = struct commod;
 struct plinfo {            /* planetary stockpiles */
   unsigned short fuel;     /* fuel for powering things */
   unsigned short destruct; /* destructive potential */
-  unsigned short resource; /* resources in storage */
+  unsigned long resource;  /* resources in storage */
   population_t popn;
   population_t troops;
   unsigned short crystals;
